src/mod-4-multi-agent/requirements.txt
# Core CrewAI Framework
crewai==0.119.0
crewai-tools==0.62.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.35  # For better ORM support if needed

# Scheduling
apscheduler==3.10.4

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Testing & Data Generation
faker==28.4.1

# Observability & Monitoring
langfuse==2.58.2
openlit==1.28.1
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Logging
colorlog==6.8.2  # Colored logging for better debugging